feat: moved api keys page
